# Fix standard deviation gradient