--- a/tools/autograd/templates/Functions.cpp
+++ b/tools/autograd/templates/Functions.cpp
@@ -580,7 +580,7 @@ Tensor std_backward(const Tensor & grad, const Tensor & self, Tensor result, boo
   if (result_zero_mask.any().toCByte()) {
     auto N = self.numel();
     double grad_zero_squared = unbiased ? 1 / N : (N - 1) / (N * N);
-    Scalar grad_zero = Scalar(grad_zero_squared).sqrt();
+    Scalar grad_zero = Scalar(grad_zero_squared).pow(0.5);
     return var_backward(grad / (2 * result), self, unbiased).masked_fill_(result_zero_mask, grad_zero);
   }
   return var_backward(grad / (2 * result), self, unbiased);
@@ -591,8 +591,8 @@ Tensor std_backward(Tensor grad, const Tensor & self, Tensor result, int64_t dim
   if (result_zero_mask.any().toCByte()) {
     auto N = self.size(dim);
     double grad_zero_squared = unbiased ? 1 / N : (N - 1) / (N * N);
-    Scalar grad_zero = Scalar(grad_zero_squared).sqrt();
-    return var_backward(grad / (2 * result), self, dim, unbiased, keepdim).masked_fill_(result == 0., grad_zero);
+    Scalar grad_zero = Scalar(grad_zero_squared).pow(0.5);
+    return var_backward(grad / (2 * result), self, dim, unbiased, keepdim).masked_fill_(result_zero_mask, grad_zero);
   }
   return var_backward(grad / (2 * result), self, dim, unbiased, keepdim);
 }