--- a/tools/autograd/templates/Functions.cpp
+++ b/tools/autograd/templates/Functions.cpp
@@ -578,9 +578,9 @@ Tensor var_backward(Tensor grad, const Tensor & self, int64_t dim, bool unbiased
 Tensor std_backward(const Tensor & grad, const Tensor & self, Tensor result, bool unbiased) {
   Tensor result_zero_mask = (result == 0.);
   if (result_zero_mask.any().toCByte()) {
-    auto N = self.numel();
-    double grad_zero_squared = unbiased ? 1 / N : (N - 1) / (N * N);
-    Scalar grad_zero = Scalar(grad_zero_squared).pow(0.5);
+    auto N = self.numel();  
+    double grad_zero_squared = unbiased ? (1. / N) : (N - 1.) / (N * N);
+    Scalar grad_zero = Scalar(sqrt(grad_zero_squared));
     return var_backward(grad / (2 * result), self, unbiased).masked_fill_(result_zero_mask, grad_zero);
   }
   return var_backward(grad / (2 * result), self, unbiased);
@@ -590,8 +590,8 @@ Tensor std_backward(Tensor grad, const Tensor & self, Tensor result, int64_t dim
   Tensor result_zero_mask = (result == 0.);
   if (result_zero_mask.any().toCByte()) {
     auto N = self.size(dim);
-    double grad_zero_squared = unbiased ? 1 / N : (N - 1) / (N * N);
-    Scalar grad_zero = Scalar(grad_zero_squared).pow(0.5);
+    double grad_zero_squared = unbiased ? 1. / N : (N - 1.) / (N * N);
+    Scalar grad_zero = Scalar(sqrt(grad_zero_squared));
     return var_backward(grad / (2 * result), self, dim, unbiased, keepdim).masked_fill_(result_zero_mask, grad_zero);
   }
   return var_backward(grad / (2 * result), self, dim, unbiased, keepdim);