Defined the subgradient of `std` when `result = 0` to be `inf`.

1. Added tests in test_autograd
2. New method std_backward() for computing backward in Functions.cpp

Fixes #4320 .