@vishwakftw 2018-07-07
I can make it one extra op alone, by modifying the entry in `Declarations.cwrap` alone.