@vishwakftw 2018-07-09
This is the time difference:
```
# Without masked_fill_(result == 0.0, INFINITY)
81 ms ± 1.71 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# With masked_fill_(result == 0.0, INFINITY)
128 ms ± 2.65 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# With recent version (optimized masked_fill)
109 ms ± 1.87 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Regarding use of `inf` instead of of any other value in `-inf, inf`, I did this to match the behaviour of `sqrt`.