@colesbury 2018-07-12
I think the formula should be:

unbiased: `1/sqrt(n)`
biased: `sqrt(n-1)/n`

You can verify this numerically

```
N = 7
eps = 1e-9
x = torch.zeros(N, dtype=torch.double)
x[0] = eps
print(float(torch.std(x) / eps))
print(float(1/math.sqrt(N)))
```