@SsnL 2018-07-12
@colesbury then you can apply the same logic and say that they should be
negated ones of what you have.

On Thu, Jul 12, 2018 at 02:04 Sam Gross <notifications@github.com> wrote:

> I think the formula should be:
>
> unbiased: 1/sqrt(n)
> biased: sqrt(n-1)/n
>
> You can verify this numerically
>
> N = 7
> eps = 1e-9
> x = torch.zeros(N, dtype=torch.double)
> x[0] = eps
> print(float(torch.std(x) / eps))
> print(float(1/math.sqrt(N)))
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/pytorch/pytorch/pull/9238#issuecomment-404347701>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AFaWZVoCh0gowWJJXhr-pR7iSMPvHdesks5uFpKegaJpZM4VGY2->
> .
>