@SsnL 2018-07-12
I think it depends on which direction you take the limit.

On Thu, Jul 12, 2018 at 02:07 Vishwak Srinivasan <notifications@github.com>
wrote:

> @colesbury <https://github.com/colesbury> Thanks for the advice. I
> computed the derivative using the definition and it came out to be:
> sqrt(n-1)/nN where N = n-1 for unbiased and n otherwise.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/pytorch/pytorch/pull/9238#issuecomment-404348126>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AFaWZVvnF9kx9qJGce4AFks6txhUSTeAks5uFpMogaJpZM4VGY2->
> .
>