@SsnL 2018-07-12
Hmmm isn’t stddev symmetrical at origin?

On Thu, Jul 12, 2018 at 02:10 Vishwak Srinivasan <notifications@github.com>
wrote:

> It won't be negative, but I think it'll be higher if you take x - delta
> instead of x + delta.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/pytorch/pytorch/pull/9238#issuecomment-404348660>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AFaWZTFob3Sr47SkMYv8QESqkgPkTwVNks5uFpPigaJpZM4VGY2->
> .
>