@vishwakftw 2018-07-12
If we take `x - delta`, then derivative is `sqrt((1 + 3/n)/N)` where `N` is `n-1` if unbiased, `n` if biased.