@ezyang 2018-07-14
@SsnL is the math good now?