@SsnL 2018-07-15
Derivative is not the same if we take the limit from `x->0-` direction. The change in the output is the same, but the change in the input is negated. So the limit is negative of that taken from `x->0+` direction. You can verify this using @colesbury 's code above but with `eps=-1e-9`. So essentially anywhere in range `[-1/sqrt(N), 1/sqrt(N)]` is valid subgradient.

Since the usage is mostly gradient-based optimization like gd, it makes sense to take one of these two extremes. But it's unclear to me that which we should choose. (`nan` really isn't a bad choice in this sense.) I think if we choose one, we should state the choice in the doc.