@vishwakftw 2018-07-17
@ezyang Here are the timings you were curious about:
```
With CUDA
With unconditional masked_fill_ in std_backward() for tensor of size 1000 x 1000 x 10 filled with 1s
33.6 ms ± 173 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

With conditional masked_fill_ in std_backward() for tensor of size 1000 x 1000 x 10 filled with 1s
33.7 ms ± 275 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

With unconditional masked_fill_ in std_backward() for tensor of size 1000 x 1000 x 10 filled using randn
31.2 ms ± 111 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

With conditional masked_fill_ in std_backward() for tensor of size 1000 x 1000 x 10 filled using randn
30.8 ms ± 209 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

`conditional masked_fill_` is the current implementation.