@ailzhang 2018-07-24
@pytorchbot retest this please