@colesbury 2018-07-30
I'm a bit wary about this change. It looks like it introduces a CUDA synchronization point in result_zero_mask.any().toCByte() (as ezyang pointed out).

synchronization points may not slow down an individual call, but they make it harder to hide kernel launch latency and CPU computation in a bigger system.