@vishwakftw 2018-08-04
Are you suggesting that the `masked_fill_` be done unconditionally?